Ground/AHRS Calibration: Make sure it returns updateRaw to false after running to stop from hammering telemetry and keep OP-AHRS smooth.  Also fixed bad connection in magic waypoint.
git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@1807 ebee16cc-31ac-478f-84a7-5cbb03baadba

+++ ground/src/plugins/magicwaypoint/magicwaypointgadgetwidget.cpp
@@ -56,7 +56,7 @@ MagicWaypointGadgetWidget::MagicWaypointGadgetWidget(QWidget *parent) : QLabel(p
     connect(this, SIGNAL(positionDesiredObjectChanged(double,double)), m_magicwaypoint->widgetPosition, SLOT(updateDesiredIndicator(double,double)));
 
     // Catch changes in scale for visualization
-    connect(m_magicwaypoint->horizontalSliderScale, SIGNAL(valueChanged(int)), this, SLOT(scaleChanged(double)));
+    connect(m_magicwaypoint->horizontalSliderScale, SIGNAL(valueChanged(int)), this, SLOT(scaleChanged(int)));
 }
 
 MagicWaypointGadgetWidget::~MagicWaypointGadgetWidget()
